+++ src/index.ts
@@ -1,4 +1,5 @@
 export * from "./components/AsyncSnapshot";
 export * from "./components/RxBuilder";
 export * from "./components/PromiseBuilder";
-export * from "./hooks/useService";
+export * from "./components/RxService";
+export { Injectable } from "./metadata/Injectable"